fix: code optimization split to api , utils , page content
